Move user/security related things to UserBundle
