+++ src/Elewant/AppBundle/EventSubscriber/NewUserSubscriber.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Elewant\AppBundle\EventSubscriber;
 
-use Elewant\AppBundle\Event\UserHasRegistered;
 use Elewant\AppBundle\Repository\HerdRepository;
 use Elewant\Herding\Model\Commands\FormHerd;
+use Elewant\UserBundle\Event\UserHasRegistered;
 use Prooph\ServiceBus\CommandBus;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
