+++ src/Elewant/UserBundle/ElewantUserBundle.php
@@ -8,4 +8,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class ElewantUserBundle extends Bundle
 {
+    public function getParent() : string
+    {
+        return 'HWIOAuthBundle';
+    }
 }
