+++ src/Elewant/UserBundle/Resources/views/Connect/connect_confirm.html.twig
@@ -1,0 +1,22 @@
+{% extends 'HWIOAuthBundle::layout.html.twig' %}
+
+{% block hwi_oauth_content %}
+    <h3>{{ 'header.connecting' | trans({}, 'HWIOAuthBundle') }}</h3>
+    <div class="row">
+        <div class="span6">
+            <p>{{ 'connect.confirm.text' | trans({'%service%': service | trans({}, 'HWIOAuthBundle'), '%name%': userInformation.realName}, 'HWIOAuthBundle') }}</p>
+            {{ form_start(form, {'action': path('hwi_oauth_connect_service', {'service': service, 'key': key}), 'attr': {'class': 'fos_user_registration_register'}}) }}
+            {{ form_widget(form) }}
+            <div>
+                <button type="submit" class="btn btn-primary">{{ 'connect.confirm.submit' | trans({}, 'HWIOAuthBundle') }}</button>
+                <a href="{{ path('hwi_oauth_connect') }}" class="btn">{{ 'connect.confirm.cancel' | trans({}, 'HWIOAuthBundle') }}</a>
+            </div>
+            {{ form_end(form) }}
+        </div>
+        <div class="span6">
+            {% if userInformation.profilePicture is defined and userInformation.profilePicture is not empty %}
+                <img src="{{ userInformation.profilePicture }}" />
+            {% endif %}
+        </div>
+    </div>
+{% endblock hwi_oauth_content %}
