+++ src/Elewant/UserBundle/Resources/views/Connect/connect_success.html.twig
@@ -1,0 +1,5 @@
+{% extends 'HWIOAuthBundle::layout.html.twig' %}
+
+{% block hwi_oauth_content %}
+    <h3>{{ 'header.success' | trans({'%name%': userInformation.realName}, 'HWIOAuthBundle') }}</h3>
+{% endblock hwi_oauth_content %}
