+++ src/Elewant/UserBundle/Resources/views/Connect/registration.html.twig
@@ -1,0 +1,22 @@
+{% extends 'HWIOAuthBundle::layout.html.twig' %}
+
+{% block hwi_oauth_content %}
+    <h3>{{ 'header.register' | trans({'%name%': userInformation.realName}, 'HWIOAuthBundle') }}</h3>
+    <div class="row">
+        <div class="span6">
+            {{ form_start(form, {'action': path('hwi_oauth_connect_registration', {'key': key}), 'attr': {'class': 'hwi_oauth_registration_register'}}) }}
+            {{ form_widget(form) }}
+            <div>
+                <button type="submit" class="btn btn-primary">{{ 'connect.registration.submit'|trans({}, 'HWIOAuthBundle') }}</button>
+                <a href="{{ path('hwi_oauth_connect') }}" class="btn">{{ 'connect.registration.cancel' | trans({}, 'HWIOAuthBundle') }}</a>
+            </div>
+            {{ form_end(form) }}
+        </div>
+        <div class="span6">
+            {% if userInformation.profilePicture is not empty %}
+                <img src="{{ userInformation.profilePicture }}" />
+            {% endif %}
+        </div>
+    </div>
+
+{% endblock hwi_oauth_content %}
