+++ src/Elewant/UserBundle/Resources/views/Connect/registration_success.html.twig
@@ -1,0 +1,5 @@
+{% extends 'HWIOAuthBundle::layout.html.twig' %}
+
+{% block hwi_oauth_content %}
+    <h3>{{ 'header.registration_success' | trans({'%username%': app.user.username}, 'HWIOAuthBundle') }}</h3>
+{% endblock hwi_oauth_content %}
