+++ src/Elewant/UserBundle/Resources/views/layout.html.twig
@@ -1,0 +1,6 @@
+{% extends 'ElewantAppBundle:Layout:base.html.twig' %}
+
+{% block body %}
+    {% block hwi_oauth_content %}
+    {% endblock %}
+{% endblock %}
